requirements.txt
alembic==1.13.3
SQLAlchemy==2.0.36
pg8000==1.29.8
requests==2.27.1
PyYAML==6.0
python-dotenv==1.0.1